+++ workspace/src/services/identity/BusTicketPlatform.Identity.Application/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace BusTicketPlatform.Identity.Application;
+
+public class Class1
+{
+
+}