+++ workspace/src/services/identity/BusTicketPlatform.Identity.Domain/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace BusTicketPlatform.Identity.Domain;
+
+public class Class1
+{
+
+}