+++ workspace/src/services/identity/BusTicketPlatform.Identity.Infrastructure/Class1.cs
@@ -1,0 +1,6 @@
+﻿namespace BusTicketPlatform.Identity.Infrastructure;
+
+public class Class1
+{
+
+}